@@ -58,7 +58,8 @@ mesheryctl app view --all
 	RunE: func(cmd *cobra.Command, args []string) error {
 		mctlCfg, err := config.GetMesheryCtl(viper.GetViper())
 		if err != nil {
-			return errors.Wrap(err, "error processing config")
+			utils.Log.Error(err)
+			return nil
 		}
 
 		application := ""
@@ -71,7 +72,8 @@ mesheryctl app view --all
 			}
 			applicationID, isID, err = utils.ValidId(args[0], "application")
 			if err != nil {
-				return err
+				utils.Log.Error(err)
+				return nil
 			}
 		}
 		var req *http.Request
@@ -95,23 +97,27 @@ mesheryctl app view --all
 
 		req, err = utils.NewRequest("GET", url, nil)
 		if err != nil {
-			return err
+			utils.Log.Error(err)
+			return nil
 		}
 
 		res, err := utils.MakeRequest(req)
 		if err != nil {
-			return err
+			utils.Log.Error(err)
+			return nil
 		}
 
 		defer res.Body.Close()
 		body, err := io.ReadAll(res.Body)
 		if err != nil {
-			return err
+			utils.Log.Error(utils.ErrReadResponseBody(err))
+			return nil
 		}
 
 		var dat map[string]interface{}
 		if err = json.Unmarshal(body, &dat); err != nil {
-			return errors.Wrap(err, "failed to unmarshal response body")
+			utils.Log.Error(utils.ErrUnmarshal(errors.Wrap(err, "failed to unmarshal response body")))
+			return nil
 		}
 		if isID {
 			if body, err = json.MarshalIndent(dat, "", "  "); err != nil {
@@ -123,10 +129,12 @@ mesheryctl app view --all
 			}
 		} else {
 			if err = json.Unmarshal(body, &response); err != nil {
-				return errors.Wrap(err, "failed to unmarshal response body")
+				utils.Log.Error(utils.ErrUnmarshal(err))
+				return nil
 			}
 			if response.TotalCount == 0 {
-				return errors.New("application does not exit. Please get an app name and try again. Use `mesheryctl app list` to see a list of applications")
+				utils.Log.Error(utils.ErrNotFound(errors.New("application does not exit. Please get an app name and try again. Use `mesheryctl app list` to see a list of applications")))
+				return nil
 			}
 			// Manage more than one apps with similar name
 			for _, app := range response.Applications {
