@@ -10,6 +10,7 @@ import (
 	"github.com/sirupsen/logrus"
 	"github.com/spf13/pflag"
 	"github.com/spf13/viper"
+	"github.com/stretchr/testify/assert"
 )
 
 var b *bytes.Buffer
@@ -81,26 +82,8 @@ func TestViewCmd(t *testing.T) {
 
 			actualResponse := b.String()
 			expectedResponse := tt.ExpectedResponse
-
-			if expectedResponse != actualResponse {
-				t.Errorf("expected response %v and actual response %v don't match", expectedResponse, actualResponse)
-			}
+			assert.Equal(t, expectedResponse, actualResponse)
 			BreakupFunc()
 		})
 	}
-}
-func TestRunChannelWithNoCmdOrFlag(t *testing.T) {
-	SetupFunc()
-	SystemCmd.SetArgs([]string{"channel"})
-	err = SystemCmd.Execute()
-
-	actualResponse := err.Error()
-
-	expectedResponse := "please specify a flag or subcommand. Use 'mesheryctl system channel --help' to display user guide." + "\n\n"
-	expectedResponse += "See https://docs.meshery.io/reference/mesheryctl/system/channel for usage details\n"
-
-	if expectedResponse != actualResponse {
-		t.Errorf("expected response %v and actual response %v don't match", expectedResponse, actualResponse)
-	}
-	BreakupFunc()
 }
