@@ -40,7 +40,10 @@ The authentication mode is web-based browser flow`,
 	Args: cobra.MinimumNArgs(0),
 	Example: `
 // Login with the Meshery Provider of your choice: the Local Provider or a Remote Provider.
-mesheryctl system login 
+mesheryctl system login
+
+// Login with the Meshery Provider by specifying it via -p or --provider flag.
+mesheryctl system login -p Meshery
 	`,
 	RunE: func(cmd *cobra.Command, args []string) error {
 		mctlCfg, err := config.GetMesheryCtl(viper.GetViper())
