@@ -497,17 +497,17 @@ func GetID(configuration string) ([]string, error) {
 	defer res.Body.Close()
 	body, err := io.ReadAll(res.Body)
 	if err != nil {
-		return idList, err
+		return idList, ErrReadResponseBody(err)
 	}
 	var dat map[string]interface{}
 	if err = json.Unmarshal(body, &dat); err != nil {
-		return idList, errors.Wrap(err, "failed to unmarshal response body")
+		return idList, ErrUnmarshal(errors.Wrap(err, "failed to unmarshal response body"))
 	}
 	if dat == nil {
-		return idList, errors.New("no data found")
+		return idList, ErrNotFound(errors.New("no data found"))
 	}
 	if dat[configType] == nil {
-		return idList, errors.New("no results found")
+		return idList, ErrNotFound(errors.New("no results found"))
 	}
 	for _, config := range dat[configType].([]interface{}) {
 		idList = append(idList, config.(map[string]interface{})["id"].(string))
@@ -533,17 +533,17 @@ func GetName(configuration string) (map[string]string, error) {
 	defer res.Body.Close()
 	body, err := io.ReadAll(res.Body)
 	if err != nil {
-		return nameIdMap, err
+		return nameIdMap, ErrReadResponseBody(err)
 	}
 	var dat map[string]interface{}
 	if err = json.Unmarshal(body, &dat); err != nil {
-		return nameIdMap, errors.Wrap(err, "failed to unmarshal response body")
+		return nameIdMap, ErrUnmarshal(errors.Wrap(err, "failed to unmarshal response body"))
 	}
 	if dat == nil {
-		return nameIdMap, errors.New("no data found")
+		return nameIdMap, ErrNotFound(errors.New("no data found"))
 	}
 	if dat[configType] == nil {
-		return nameIdMap, errors.New("no results found")
+		return nameIdMap, ErrNotFound(errors.New("no results found"))
 	}
 	for _, config := range dat[configType].([]interface{}) {
 		nameIdMap[config.(map[string]interface{})["name"].(string)] = config.(map[string]interface{})["id"].(string)
@@ -576,10 +576,12 @@ func ValidId(args string, configuration string) (string, bool, error) {
 				args = id
 			}
 		}
+	} else {
+		return "", false, err
 	}
 	isID, err = regexp.MatchString("^[a-fA-F0-9]{8}-[a-fA-F0-9]{4}-4[a-fA-F0-9]{3}-[8|9|aA|bB][a-fA-F0-9]{3}-[a-fA-F0-9]{12}$", args)
 	if err != nil {
-		return "", false, err
+		return "", false, ErrInvalidNameOrID(err)
 	}
 	return args, isID, nil
 }
@@ -635,7 +637,7 @@ func ParseURLGithub(URL string) (string, string, error) {
 	// - https://raw.githubusercontent.com/layer5io/meshery/master/.goreleaser.yml
 	parsedURL, err := url.Parse(URL)
 	if err != nil {
-		return "", "", fmt.Errorf("failed to retrieve file from URL: %s", URL)
+		return "", "", ErrParsingUrl(fmt.Errorf("failed to retrieve file from URL: %s", URL))
 	}
 	host := parsedURL.Host
 	path := parsedURL.Path
@@ -643,18 +645,18 @@ func ParseURLGithub(URL string) (string, string, error) {
 	paths := strings.Split(path, "/")
 	if host == "github.com" {
 		if len(paths) < 5 {
-			return "", "", fmt.Errorf("failed to retrieve file from URL: %s", URL)
+			return "", "", ErrParsingUrl(fmt.Errorf("failed to retrieve file from URL: %s", URL))
 		}
 		resURL := "https://" + host + strings.Join(paths[:4], "/")
 		return resURL, strings.Join(paths[4:], "/"), nil
 	} else if host == "raw.githubusercontent.com" {
 		if len(paths) < 5 {
-			return "", "", fmt.Errorf("failed to retrieve file from URL: %s", URL)
+			return "", "", ErrParsingUrl(fmt.Errorf("failed to retrieve file from URL: %s", URL))
 		}
 		resURL := "https://" + "raw.githubusercontent.com" + path
 		return resURL, "", nil
 	}
-	return URL, "", errors.New("only github urls are supported")
+	return URL, "", ErrParsingUrl(errors.New("only github urls are supported"))
 }
 
 // PrintToTableInStringFormat prints the given data into a table format but return as a string
@@ -679,6 +681,7 @@ func PrintToTableInStringFormat(header []string, data [][]string) string {
 	return tableString.String()
 }
 
+// Indicate an ongoing Process at a given time on CLI
 func CreateDefaultSpinner(suffix string, finalMsg string) *spinner.Spinner {
 	s := spinner.New(spinner.CharSets[11], 100*time.Millisecond)
 
@@ -687,6 +690,7 @@ func CreateDefaultSpinner(suffix string, finalMsg string) *spinner.Spinner {
 	return s
 }
 
+// Get Meshery Session Data/Details (Adapters)
 func GetSessionData(mctlCfg *config.MesheryCtlConfig) (*models.Preference, error) {
 	path := mctlCfg.GetBaseMesheryURL() + "/api/system/sync"
 	method := "GET"
