@@ -145,12 +145,12 @@ func K8sContextsFromKubeconfig(kubeconfig []byte, instanceID *uuid.UUID) ([]*K8s
 
 	parsed, err := clientcmd.Load(kubeconfig)
 	if err != nil {
-		return kcs, respMessage
+		return kcs, err.Error()
 	}
 
 	kcfg := InternalKubeConfig{}
 	if err := yaml.Unmarshal(kubeconfig, &kcfg); err != nil {
-		return kcs, respMessage
+		return kcs, err.Error()
 	}
 
 	for name := range parsed.Contexts {
@@ -163,15 +163,25 @@ func K8sContextsFromKubeconfig(kubeconfig []byte, instanceID *uuid.UUID) ([]*K8s
 			respMessage += msg
 			continue
 		}
+		
+		// Perform Ping test on the cluster
+		if err := kc.PingTest(); err != nil {
+			msg = fmt.Sprintf("Skipping context: %v \n", err)
+			logrus.Warn(msg)
+			respMessage += msg
+			continue
+		}
+
 		err = kc.AssignVersion(handler)
 		if err != nil {
-			msg = fmt.Sprintf("error getting kubernetes version: %v\n Skipping context", err)
+			msg = fmt.Sprintf("Skipping context: Could not retrieve Kubernetes version: %v ", err)
 			logrus.Warnf(msg)
 			respMessage += msg
 			continue
 		}
+		
 		if err := kc.AssignServerID(handler); err != nil {
-			msg = fmt.Sprintf("Skipping context: Reason => %s\n", err)
+			msg = fmt.Sprintf("Skipping context: Could not retrieve Kubernetes cluster ID (%v)", err)
 			logrus.Warn(msg)
 			respMessage += msg
 			continue
@@ -345,11 +355,6 @@ func (kc K8sContext) PingTest() error {
 // AssignServerID will attempt to assign kubernetes
 // server ID to the kubernetes context
 func (kc *K8sContext) AssignServerID(handler *kubernetes.Client) error {
-	// Perform Ping test on the cluster
-	if err := kc.PingTest(); err != nil {
-		return err
-	}
-
 	// Get Kubernetes API server ID by querying the "kube-system" namespace uuid
 	ksns, err := handler.KubeClient.CoreV1().Namespaces().Get(context.TODO(), "kube-system", v1.GetOptions{})
 	if err != nil {
