@@ -10,18 +10,18 @@ import {
   Typography,
   Tabs,
   Tab,
-  ClickAwayListener
+  ClickAwayListener,
 } from '@material-ui/core';
 import BellIcon from '@material-ui/icons/Notifications';
 import ClearIcon from "../assets/icons/ClearIcon";
 import ErrorIcon from '@material-ui/icons/Error';
 import { withStyles } from '@material-ui/core/styles';
 import amber from '@material-ui/core/colors/amber';
-import {  EVENT_TYPES, SERVER_EVENT_TYPES } from '../lib/event-types';
-import MesheryEventViewer from './MesheryEventViewer';
+import {  EVENT_TYPES, NOTIFICATION_STATUS, SERVER_EVENT_TYPES } from '../lib/event-types';
+import Notification from "./NotificationCenter/Notification"
 import dataFetch from '../lib/data-fetch';
 import { bindActionCreators } from "redux";
-import { toggleNotificationCenter, updateEvents } from "../lib/store";
+import { loadEventsFromPersistence, toggleNotificationCenter, updateEvents } from "../lib/store";
 import { iconMedium } from "../css/icons.styles";
 import { cursorNotAllowed } from "../css/disableComponent.styles";
 import { v4 } from "uuid";
@@ -29,7 +29,7 @@ import moment from "moment";
 import { withNotify } from "../utils/hooks/useNotification";
 
 const styles = (theme) => ({
-  sidelist : { width : 450, },
+  sidelist : { width : "35rem", },
   notificationButton : { height : '100%', },
   notificationDrawer : {
     backgroundColor : theme.palette.secondary.sideBar,
@@ -80,22 +80,42 @@ const styles = (theme) => ({
     transition : '0.3s ease-in-out !important'
   },
   peekView : {
-    right : "-26.1rem",
+    right : "-32.1rem",
     transition : '0.3s ease-in-out !important'
   },
   tabs : {
+
+    "&.MuiTabs-flexContainer" : {
+      display : "flex",
+      justifyContent : "space-around",
+    },
     "& .MuiTabs-indicator" : {
       backgroundColor : theme.palette.type === 'dark' ? "#00B39F" : theme.palette.primary,
     },
   },
   tab : {
+    padding : "0px",
+    margin : "0px",
     "&.Mui-selected" : {
       color : theme.palette.type === 'dark' ? "#00B39F" : theme.palette.primary,
     },
+  },
+
+  notification : {
+    margin : theme.spacing(0.5, 1)
   }
 });
 
 
+
+
+export const NOTIFICATION_FILTERS = {
+  ALL : "all",
+  ERROR : EVENT_TYPES.ERROR.type ,
+  SUCCESS : EVENT_TYPES.SUCCESS.type ,
+  WARNING : EVENT_TYPES.WARNING.type,
+  HISTORY : "history" ,
+}
 
 /**
  * getNotifications filters the notifications based on the
@@ -114,16 +134,20 @@ const styles = (theme) => ({
  *  operation_id: string
  * }[]}
  */
-function getNotifications(events, type) {
+function getNotifications(events, filter) {
   if (!Array.isArray(events)) return [];
+  if (filter == NOTIFICATION_FILTERS.HISTORY) return events.filter(ev => ev.status === NOTIFICATION_STATUS.VIEWED)
 
-  if (type === "error") return events.filter(ev => getEventType(ev).type == EVENT_TYPES.ERROR.type);
-  if (type === "warning") return events.filter(ev => getEventType(ev).type == EVENT_TYPES.WARNING.type)
-  if (type === "success") return events.filter(ev =>
-    ((getEventType(ev).type == EVENT_TYPES.SUCCESS.type)
-                                      || getEventType(ev).type == EVENT_TYPES.INFO.type))
+  if (filter == NOTIFICATION_FILTERS.ALL) return events.filter(ev => ev.status !== NOTIFICATION_STATUS.VIEWED)
 
-  return events;
+  if (filter === NOTIFICATION_FILTERS.SUCCESS) {
+    return events.filter(ev => {
+      const ev_type = getEventType(ev).type
+      return (ev_type == (EVENT_TYPES.SUCCESS.type || ev_type == EVENT_TYPES.INFO.type) &&  ev.status !== NOTIFICATION_STATUS.VIEWED )
+    })
+  }
+  return events.filter(ev => getEventType(ev).type == filter && ev.status !== NOTIFICATION_STATUS.VIEWED )
+
 }
 
 /**
@@ -138,8 +162,8 @@ function getNotifications(events, type) {
 function getNotificationCount(events) {
   if (!Array.isArray(events)) return 0;
 
-  const errorEventCount = events.filter(ev => getEventType(ev).type == EVENT_TYPES.ERROR.type ).length;
-  const totalEventsCount = events.length;
+  const errorEventCount = getNotifications(events,NOTIFICATION_FILTERS.ERROR).length;
+  const totalEventsCount = getNotifications(events,NOTIFICATION_FILTERS.ALL).length;
   return errorEventCount || totalEventsCount;
 }
 
@@ -164,19 +188,35 @@ function NotificationIcon({ type, className }) {
   return <BellIcon className={className} style={iconMedium} />
 }
 
+const notificationBadgeTooltipMessage = (events) => {
+  const total_unread = getNotifications(events,NOTIFICATION_FILTERS.ALL).length
+  if (total_unread) {
+    return `${total_unread} new notifications`
+  }
+  return `No new notifications`
+}
+
+const TabLabel = ({ filterType,events }) => {
+
+  const notifCount = getNotifications(events,filterType).length
+  return (
+    <div style={{ margin : 0 , padding : 0 ,textTransform : "capitalize" , display : "flex" , gap : "0.3rem" }}>
+      {notifCount ? `${filterType} (${notifCount})` : filterType }
+    </div>
+  )
+}
+
 //TODO: Convert To functional Compoent
 class MesheryNotification extends React.Component {
   state = {
     open : false,
     dialogShow : false,
-    displayEventType : "*",
+    displayEventType : NOTIFICATION_FILTERS.ALL,
     tabValue : 0,
     anchorEl : false,
-    // showFullNotificationCenter : false,
   }
 
   handleToggle = () => {
-    // this.setState({ showFullNotificationCenter : !this.state.showFullNotificationCenter })
     this.props.toggleOpen()
   };
 
@@ -201,6 +241,12 @@ class MesheryNotification extends React.Component {
     this.startEventStream();
   }
 
+  componentDidUpdate() {
+    if (this.props.user.get("user_id") && this.props.events.length == 0) {
+      this.props.loadEventsFromPersistence()
+    }
+  }
+
   async startEventStream() {
     this.closeEventStream();
     this.eventStream = new EventSource('/api/events');
@@ -215,6 +261,7 @@ class MesheryNotification extends React.Component {
       const data = JSON.parse(e.data);
       const event = {
         ...data,
+        status : NOTIFICATION_STATUS.NEW,
         event_type : getEventType(data),
         timestamp : data.timestamp || moment.utc().valueOf() ,
         id : data.id || v4() ,
@@ -250,12 +297,10 @@ class MesheryNotification extends React.Component {
     }
   }
 
-  deleteEvent = (ind) => () => {
+  deleteEvent = (id) => {
     const { events, updateEvents } = this.props;
-    if (events[ind]) {
-      events.splice(ind, 1);
-    }
-    updateEvents({ events : events })
+    const newEvents = events.filter(ev => ev.id !== id)
+    updateEvents({ events : newEvents })
     this.setState({ dialogShow : false });
   }
 
@@ -278,7 +323,7 @@ class MesheryNotification extends React.Component {
     }
   }
 
-  handleTabChange = (event, newTabValue) => {
+  handleTabChange = (_event, newTabValue) => {
     this.setState({ tabValue : newTabValue })
   }
 
@@ -289,41 +334,33 @@ class MesheryNotification extends React.Component {
     })
   }
 
+  markAsRead = (event) => {
+    const events = this.props.events.filter(ev => ev.id !== event.id)
+    events.push({
+      ...event,
+      status : NOTIFICATION_STATUS.VIEWED
+    })
+    this.props.updateEvents({
+      events
+    })
+  }
+
+
   render() {
-    const { classes, events ,showFullNotificationCenter } = this.props;
+    const { classes, events ,showFullNotificationCenter  } = this.props;
     const { anchorEl, show } = this.state;
-    const self = this;
     let open = Boolean(anchorEl);
     if (showFullNotificationCenter) {
       open = showFullNotificationCenter;
     }
 
-
-    let toolTipMsg;
-    if (typeof events?.length !== 'undefined') {
-      toolTipMsg = `There are ${events.length} events`;
-      switch (events?.length) {
-        case 0:
-          toolTipMsg = 'There are no events';
-          break;
-        case 1:
-          toolTipMsg = 'There is 1 event';
-          break;
-      }
-    } else { // takes care of case when (typeof events.length === undefined)
-      toolTipMsg = 'There are no events';
-    }
-    let badgeColorVariant = 'default';
-    events.forEach((eev) => {
-      if ( getEventType(eev).type == EVENT_TYPES.ERROR.type ) {
-        badgeColorVariant = 'error';
-      }
-    });
+    const newErrors =  getNotifications(events,NOTIFICATION_FILTERS.ERROR)
+    const newNotificationsType = newErrors.length > 0 ? "error" : "default"
 
     return (
       <NoSsr>
         <div style={ show ? cursorNotAllowed : {}}>
-          <Tooltip title={toolTipMsg}>
+          <Tooltip title={notificationBadgeTooltipMessage(events)}>
             <IconButton
               id="notification-button"
               className={classes.notificationButton}
@@ -343,8 +380,8 @@ class MesheryNotification extends React.Component {
                 this.setState({ anchorEl : false })
               }}
             >
-              <Badge id="notification-badge" badgeContent={getNotificationCount(events)} color={badgeColorVariant}>
-                <NotificationIcon  style={iconMedium}  type={badgeColorVariant} />
+              <Badge id="notification-badge" badgeContent={getNotificationCount(events)} color={newNotificationsType}>
+                <NotificationIcon  style={iconMedium}  type={newNotificationsType}   />
               </Badge>
             </IconButton>
           </Tooltip>
@@ -407,24 +444,21 @@ class MesheryNotification extends React.Component {
                     textColor="primary"
                     variant="fullWidth"
                   >
-                    <Tab label="All" className={classes.tab} onClick={this.handleNotifFiltering('*')} style={{ minWidth : "15%" }} />
-                    <Tab label="Error" className={classes.tab} onClick={this.handleNotifFiltering('error')} style={{ minWidth : "15%" }} />
-                    <Tab label="Warning" className={classes.tab} onClick={this.handleNotifFiltering('warning')} style={{ minWidth : "15%" }} />
-                    <Tab label="Success" className={classes.tab} onClick={this.handleNotifFiltering('success')} style={{ minWidth : "15%" }} />
+
+                    {Object.values(NOTIFICATION_FILTERS).map(filter  =>
+                      <Tab label={<TabLabel filterType={filter} events={events}/>}
+                        className={classes.tab}
+                        onClick={this.handleNotifFiltering(filter)}
+                        style={{ minWidth : "8%" }} />
+                    )}
                   </Tabs>
-                  {getNotifications(this.props.events, this.state.displayEventType).map((event, ind) => (
-                    <MesheryEventViewer
-                      key={ind}
-                      eventVariant={getEventType(event)}
-                      eventSummary={event.summary}
-                      deleteEvent={self.deleteEvent(ind)}
-                      eventDetails={event.details || "Details Unavailable"}
-                      eventCause={event.probable_cause}
-                      eventRemediation={event.suggested_remediation}
-                      eventErrorCode={event.error_code}
-                      componentType={event.component}
-                      componentName={event.component_name}
-                      eventTimestamp={event.timestamp}
+                  {getNotifications(this.props.events, this.state.displayEventType).map((event) => (
+                    <Notification
+                      className = {classes.notification}
+                      key={event.id}
+                      event={event}
+                      onDeleteEvent = {() => this.deleteEvent(event.id)}
+                      onMarkAsRead = {() => this.markAsRead(event) }
                       expand={(this.props.openEventId && this.props.openEventId === ( event.id) ) ? true : false}
                     />
                   ))}
@@ -440,17 +474,14 @@ class MesheryNotification extends React.Component {
 
 const mapDispatchToProps = (dispatch) => ({
   updateEvents : bindActionCreators(updateEvents, dispatch),
-  toggleOpen : bindActionCreators(toggleNotificationCenter,dispatch)
+  toggleOpen : bindActionCreators(toggleNotificationCenter,dispatch),
+  loadEventsFromPersistence : bindActionCreators(loadEventsFromPersistence,dispatch)
 })
 
-// const mapDispatchToProps = (dispatch) => ({
-//   updateSMIResults: bindActionCreators(updateSMIResults, dispatch),
-// });
-
 const mapStateToProps = (state) => {
-  //TODO: Sort While Storing
   const events = state.get('events')
   return {
+    user : state.get("user"),
     events : events.toJS(),
     openEventId : state.get("notificationCenter").get("openEventId"),
     showFullNotificationCenter : state.get("notificationCenter").get("showFullNotificationCenter")
