@@ -13,7 +13,7 @@ module.exports = defineConfig({
     setupNodeEvents(on, config) {
       return require('./cypress/plugins/index.js')(on, config)
     },
-    baseUrl : 'http://localhost:3000',
+    baseUrl : 'http://localhost:9081',
     specPattern : 'cypress/e2e/**/*.{js,jsx,ts,tsx}',
     pageLoadTimeout : 120_000,
   }
