@@ -86,10 +86,12 @@ class MesheryApp extends App {
     super();
     this.pageContext = getPageContext();
     this.meshsyncEventsSubscriptionRef = React.createRef();
+    this.fullScreenChanged = this.fullScreenChanged.bind(this);
 
     this.state = {
       mobileOpen : false,
       isDrawerCollapsed : false,
+      isFullScreenMode : false,
       k8sContexts : [],
       activeK8sContexts : [],
       operatorSubscription : null,
@@ -138,6 +140,12 @@ class MesheryApp extends App {
     this.meshsyncEventsSubscriptionRef.current = meshSyncEventsSubscription;
   }
 
+  fullScreenChanged = () => {
+    this.setState(state => {
+      return { isFullScreenMode : !state.isFullScreenMode }
+    });
+  }
+
   componentDidMount() {
     this.loadConfigFromServer(); // this works, but sometimes other components which need data load faster than this data is obtained.
     this.initSubscriptions([]);
@@ -175,6 +183,12 @@ class MesheryApp extends App {
     }
     const disposeK8sContextSubscription = k8sContextSubscription();
     this.setState({ disposeK8sContextSubscription })
+
+    document.addEventListener("fullscreenchange", this.fullScreenChanged);
+  }
+
+  componentWillUnmount() {
+    document.removeEventListener("fullscreenchange", this.fullScreenChanged);
   }
 
   componentDidUpdate(prevProps) {
@@ -350,13 +364,14 @@ class MesheryApp extends App {
     const {
       Component, pageProps, classes, isDrawerCollapsed, relayEnvironment
     } = this.props;
+
     return (
       <RelayEnvironmentProvider environment={relayEnvironment}>
         <ThemeProvider theme={this.state.theme === "dark" ? darkTheme : theme}>
           <NoSsr>
             <div className={classes.root}>
               <CssBaseline />
-              <nav className={isDrawerCollapsed
+              {!this.state.isFullScreenMode && <nav className={isDrawerCollapsed
                 ? classes.drawerCollapsed
                 : classes.drawer} data-test="navigation">
                 <Hidden smUp implementation="js">
@@ -377,6 +392,7 @@ class MesheryApp extends App {
                   />
                 </Hidden>
               </nav>
+              }
               <div className={classes.appContent}>
                 <SnackbarProvider
                   anchorOrigin={{
@@ -398,7 +414,7 @@ class MesheryApp extends App {
                   maxSnack={10}
                 >
                   <MesheryProgressBar />
-                  <Header
+                  {!this.state.isFullScreenMode &&  <Header
                     onDrawerToggle={this.handleDrawerToggle}
                     onDrawerCollapse={isDrawerCollapsed}
                     contexts={this.state.k8sContexts}
@@ -409,6 +425,7 @@ class MesheryApp extends App {
                     theme={this.state.theme}
                     themeSetter={this.themeSetter}
                   />
+                  }
                   <main className={classes.mainContent}>
                     <MuiPickersUtilsProvider utils={MomentUtils}>
                       <Component
